Refactor UnitView components to follow our component standard
