+++ src/views/UnitView/components/ContactInfo/index.js
@@ -1,0 +1,3 @@
+import ContactInfo from './ContactInfo';
+
+export default ContactInfo;
