+++ src/views/UnitView/components/Description/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@material-ui/core';
+import Description from './Description';
+import styles from '../../styles/styles';
+
+export default withStyles(styles)(Description);
