+++ src/views/UnitView/components/ElectronicServices/index.js
@@ -1,0 +1,4 @@
+import { injectIntl } from 'react-intl';
+import ElectronicServices from './ElectronicServices';
+
+export default injectIntl(ElectronicServices);
