+++ src/views/UnitView/components/Highlights/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@material-ui/core';
+import Highlights from './Highlights';
+import styles from '../../styles/styles';
+
+export default withStyles(styles)(Highlights);
