+++ src/views/UnitView/components/SocialMediaLinks/index.js
@@ -1,0 +1,5 @@
+import { withStyles } from '@material-ui/core';
+import SocialMediaLinks from './SocialMediaLinks';
+import socialMediaLinksStyles from '../../styles/socialMediaLinksStyles';
+
+export default withStyles(socialMediaLinksStyles)(SocialMediaLinks);
