+++ src/views/UnitView/components/UnitLinks/index.js
@@ -1,0 +1,4 @@
+import { injectIntl } from 'react-intl';
+import UnitLinks from './UnitLinks';
+
+export default injectIntl(UnitLinks);
